adds a test version of a function that downloads popular movies
