+++ src/js/refs.js
@@ -2,6 +2,5 @@ export const refs = {
   galleryLibraryEl: document.querySelector('.gallery-library'),
   searchMovieForm: document.querySelector('.search-movie-form'),
   sectionContainer: document.querySelector('.movies-section-container'),
-  dayBtn: document.querySelector('.day'),
-  weekBtn: document.querySelector('.week'),
+  moviesList: document.querySelector('.js-movies-list')
 };
