+++ src/partials/movies.html
@@ -1,7 +1,7 @@
 <section class="movies">
   <div class="container movies-section-container">
     <h1 class="visually-hidden">Movie library</h1>
-    <ul class="gallery-library movies-list"></ul>
+    <ul class="list gallery-library movies-list js-movies-list"></ul>
     <!-- строка 4 скорее всего - будет не нужно в будущем-->
   </div>
 </section>
