+++ src/templates/home-hopular-movies-nav.hbs
@@ -1,0 +1,9 @@
+<h2 class="popular-movies-title">Popular movies</h2>
+<ul class="list popular-btns-list">
+  <li class="popular-btns-item">
+    <button type="button" class="popular-btn day">Day</button>
+  </li>
+  <li>
+    <button type="button" class="popular-btn week">Week</button>
+  </li class="popular-btns-item">
+</ul>
